More robust fix to allowing URLs to be passed into uncss
Unless `nonull: true` is set on file object, Grunt automatically
removes paths which it can’t find locally, preventing us from passing
remote paths to the plugin and to `uncss`.

The previous workaround was to use `file.orig.src` which still has
those paths, but that property fails to include multiple paths on a
single file object and seems to always be a string, not an array, so it
fails with `file.orig.src.forEach` not a function.

Additionally, `uncss` doesn't use `options.urls` so that has been
removed.